Move port position getting logic into port
