+++ lib/model/edge.js
@@ -39,7 +39,7 @@ var Edge = createClass({
     getStartPositionIn: function(graph) {
       var port = this.get('from');
       
-      return this._getPortPositionIn(port, graph);
+      return port.getPositionInGraph(graph);
     },
 
     /**
@@ -56,18 +56,9 @@ var Edge = createClass({
     getEndPositionIn: function(graph) {
       var port = this.get('to');
       
-      return this._getPortPositionIn(port, graph);
-    },
-
-    _getPortPositionIn: function(port, graph) {
-      var vertexId = port.get('ownerVertexId');
-      var vertex = graph.getVertexById(vertexId);
-      if (vertex === null) {
-        return null;
-      }
-
-      return vertex.getPortPositionInContext(port.get('position'), true);
+      return port.getPositionInGraph(graph);
     }
+    
   },
 
   attrs: {
