+++ lib/model/port.js
@@ -3,6 +3,7 @@ var AttrValidators = require('./../utilities/attrValidators');
 var createClass = require('./../utilities/createClass');
 var GraphModelBase = require('./graphModelBase');
 var PortType = require('./../enum/portType');
+var PositionUtils = require('../utilities/positionUtils.js');
 
 var CardinalPortPosition = require('./cardinalPortPosition');
 
@@ -18,6 +19,43 @@ var CardinalPortPosition = require('./cardinalPortPosition');
 
 var Port = createClass({
   extend: GraphModelBase,
+
+  instance: {
+
+    /**
+     * Get cartesian coordinates of a port as if it were in this node.
+     * 
+     * Returns null if the vertex owner is not in this graph.
+     * (Maybe it's in in a sub- or parent container?)
+     * 
+     * @method getPortPositionInContext
+     * @param  {CardinalPortPosition} portPosition
+     * @param  {Boolean} [useGraphSpace] If true, coordinates are in the container's coordinate space instead of the node's
+     * @return {Object} With properties x, y, angle (radians)
+     */
+    getPositionInGraph: function(graph) {
+      var portPosition = this.get('position');
+      var cartesianPos = PositionUtils.Conversion.cardinalToCartesian(portPosition);
+
+      var vertexId = this.get('ownerVertexId');
+      var vertex = graph.getVertexById(vertexId);
+      if (vertex === null) {
+        return null;
+      }
+
+      var vertexStyles = vertex.get('styles');
+      var vPos = vertex.get('position');
+
+      var result = {
+        x: cartesianPos.getX() * vertexStyles.get('width') + vPos.x,
+        y: cartesianPos.getY() * vertexStyles.get('height') + vPos.y,
+        angle: portPosition.getAngle()
+      };
+
+      return result;
+    }
+  },
+
   attrs: {
     ownerVertexId: {
       validator: Lang.isString,
