+++ lib/model/vertex.js
@@ -5,7 +5,7 @@ var GraphModelBase = require('./graphModelBase');
 var PortType = require('./../enum/portType');
 var PositionUtils = require('../utilities/positionUtils.js');
 var VertexStyles = require('../styles/vertexStyles');
-var AttrValidators = require('../styles/attrValidators');
+var AttrValidators = require('../utilities/attrValidators');
 
 var Port = require('./port');
 
@@ -51,35 +51,6 @@ var Vertex = createClass({
     this.get('outputs').forEach(function(port) {
       port.addBubbleTarget(this);
     }.bind(this));
-  },
-
-  instance: {
-
-    /**
-     * Get cartesian coordinates of a port as if it were in this node.
-     * @method getPortPositionInContext
-     * @param  {CardinalPortPosition} portPosition
-     * @param  {Boolean} [useGraphSpace] If true, coordinates are in the container's coordinate space instead of the node's
-     * @return {Object} With properties x, y, angle (radians)
-     */
-    getPortPositionInContext: function(portPosition, useGraphSpace) {
-      var cartesianPos = PositionUtils.Conversion.cardinalToCartesian(portPosition);
-      var styles = this.get('styles');
-
-      var result = {
-        x: cartesianPos.getX() * styles.get('width'),
-        y: cartesianPos.getY() * styles.get('height'),
-        angle: portPosition.getAngle()
-      };
-
-      if (useGraphSpace) {
-        var myPos = this.get('position');
-        result.x += myPos.x;
-        result.y += myPos.y;
-      }
-
-      return result;
-    }
   },
 
   attrs: {
